safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-svgdom-limitingConeAngle-prop-expected.txt

CONSOLE MESSAGE: line 44: TypeError: 'undefined' is not an object (evaluating 'spotLightElement.limitingConeAngle.baseVal = 15')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'limitingConeAngle' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL spotLightElement.limitingConeAngle.baseVal should be 45. Threw exception TypeError: 'undefined' is not an object (evaluating 'spotLightElement.limitingConeAngle.baseVal')
